// src/services/sited_io/commerce/v1/shop_pb.ts
export class ShopResponse extends Message<ShopResponse> {
  /**
   * @generated from field: string shop_id = 1;
   */
  shopId = "";

  /**
   * @generated from field: string user_id = 2;
   */
  userId = "";

  /**
   * @generated from field: uint64 created_at = 3;
   */
  createdAt = protoInt64.zero;

  /**
   * @generated from field: uint64 updated_at = 4;
   */
  updatedAt = protoInt64.zero;

  /**
   * @generated from field: string name = 5;
   */
  name = "";

  /**
   * @generated from field: string slug = 6;
   */
  slug = "";

  /**
   * @generated from field: optional string domain = 7;
   */
  domain?: string;

  /**
   * @generated from field: optional string description = 8;
   */
  description?: string;

  /**
   * @generated from field: uint32 platform_fee_percent = 9;
   */
  platformFeePercent = 0;

  /**
   * @generated from field: uint32 minimum_platform_fee_cent = 10;
   */
  minimumPlatformFeeCent = 0;

  /**
   * @generated from field: optional sited_io.commerce.v1.ShopCustomizationResponse customization = 11;
   */
  customization?: ShopCustomizationResponse;

  /**
   * @generated from field: bool is_active = 12;
   */
  isActive = false;

  /**
   * @generated from field: optional string contact_email_address = 13;
   */
  contactEmailAddress?: string;

  /**
   * @generated from field: optional string client_id = 14;
   */
  clientId?: string;

  /**
   * @generated from field: string website_id = 15;
   */
  websiteId = "";

  constructor(data?: PartialMessage<ShopResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v1.ShopResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "shop_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "user_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "created_at", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 4, name: "updated_at", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
    { no: 5, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 6, name: "slug", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    {
      no: 7,
      name: "domain",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
      opt: true,
    },
    {
      no: 8,
      name: "description",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
      opt: true,
    },
    {
      no: 9,
      name: "platform_fee_percent",
      kind: "scalar",
      T: 13 /* ScalarType.UINT32 */,
    },
    {
      no: 10,
      name: "minimum_platform_fee_cent",
      kind: "scalar",
      T: 13 /* ScalarType.UINT32 */,
    },
    {
      no: 11,
      name: "customization",
      kind: "message",
      T: ShopCustomizationResponse,
      opt: true,
    },
    { no: 12, name: "is_active", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    {
      no: 13,
      name: "contact_email_address",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
      opt: true,
    },
    {
      no: 14,
      name: "client_id",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
      opt: true,
    },
    {
      no: 15,
      name: "website_id",
      kind: "scalar",
      T: 9 /* ScalarType.STRING */,
    },
  ]);

  static fromBinary(
    bytes: Uint8Array,
    options?: Partial<BinaryReadOptions>,
  ): ShopResponse {
    return new ShopResponse().fromBinary(bytes, options);
  }

  static fromJson(
    jsonValue: JsonValue,
    options?: Partial<JsonReadOptions>,
  ): ShopResponse {
    return new ShopResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(
    jsonString: string,
    options?: Partial<JsonReadOptions>,
  ): ShopResponse {
    return new ShopResponse().fromJsonString(jsonString, options);
  }

  static equals(
    a: ShopResponse | PlainMessage<ShopResponse> | undefined,
    b: ShopResponse | PlainMessage<ShopResponse> | undefined,
  ): boolean {
    return proto3.util.equals(ShopResponse, a, b);
  }
}